# expression: implement vectorized evaluation for `builtinCoalesceXXSig`