--- a/expression/generator/compare_vec.go
+++ b/expression/generator/compare_vec.go
@@ -87,8 +87,6 @@ func (b *builtin{{ .compare.CompareName }}{{ .type.TypeName }}Sig) vecEvalInt(in
 		}
 {{- if eq .type.ETName "Json" }}
 		val := json.CompareBinary(buf0.GetJSON(i), buf1.GetJSON(i))
-{{ else if eq .type.ETName "Int" }}
-		val := types.CompareInt64(arg0[i],arg1[i])
 {{ else if eq .type.ETName "Real" }}
 		val := types.CompareFloat64(arg0[i], arg1[i])
 {{- else if eq .type.ETName "String" }}
@@ -150,9 +148,7 @@ func (b *builtin{{ .compare.CompareName }}{{ .type.TypeName }}Sig) vecEvalInt(in
 			i64s[i] = 0
 {{- if eq .type.ETName "Json" }}
 		case json.CompareBinary(buf0.GetJSON(i), buf1.GetJSON(i)) == 0:
-{{ else if eq .type.ETName "Int" }}
-		case types.CompareInt64(arg0[i], arg1[i]) == 0:
-{{ else if eq .type.ETName "Real" }}
+{{- else if eq .type.ETName "Real" }}
 		case types.CompareFloat64(arg0[i], arg1[i]) == 0:
 {{- else if eq .type.ETName "String" }}
 		case types.CompareString(buf0.GetString(i), buf1.GetString(i)) == 0: