… expression/builtin_compare_vec_generated_test.go
… expression/builtin_compare_vec_test.go