<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

use go generate to implement vectorized evaluation for Coalesce
### What is changed and how it works?
```
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceIntSig-VecBuiltinFunc-12              	  189350	      6218 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceIntSig-NonVecBuiltinFunc-12           	   59558	     19551 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceRealSig-VecBuiltinFunc-12             	  190057	      6094 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceRealSig-NonVecBuiltinFunc-12          	   61264	     18935 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceDecimalSig-VecBuiltinFunc-12          	  106948	     10088 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceDecimalSig-NonVecBuiltinFunc-12       	   52993	     22329 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceStringSig-VecBuiltinFunc-12           	   81129	     14239 ns/op	      32 B/op	       1 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceStringSig-NonVecBuiltinFunc-12        	   40249	     29148 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceTimeSig-VecBuiltinFunc-12             	  131838	      8231 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceTimeSig-NonVecBuiltinFunc-12          	   31866	     37015 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceDurationSig-VecBuiltinFunc-12         	  183813	      6232 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceDurationSig-NonVecBuiltinFunc-12      	   62823	     18709 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceJSONSig-VecBuiltinFunc-12             	   81266	     14039 ns/op	      32 B/op	       1 allocs/op
BenchmarkVectorizedGeneratedBuiltinCompareFunc/builtinCoalesceJSONSig-NonVecBuiltinFunc-12          	   45518	     25790 ns/op	       0 B/op	       0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test