@SunRunAway 2019-10-30
Hi @gtygo ,
Thanks for your pull request, you should commit both the generated code and generator.
And since #12875 is merged, you may need to combine your generator/compare_vec.go into his, then run `go generate` again, so that the conflicts in the generated files can be resolved.