@gtygo 2019-11-03
hi @SunRunAway , I  regenerated (LT/LE/GT/GE/EQ/NE/NullEQ Int)  by go generate.